+++ compiler/src/org.graalvm.compiler.core/src/org/graalvm/compiler/core/target/Backend.java
@@ -31,6 +31,7 @@ import org.graalvm.compiler.core.common.CompilationIdentifier;
 import org.graalvm.compiler.core.common.LIRKind;
 import org.graalvm.compiler.core.common.alloc.BasicBlockOrder;
 import org.graalvm.compiler.core.common.alloc.RegisterAllocationConfig;
+import org.graalvm.compiler.core.common.cfg.AbstractBlockBase;
 import org.graalvm.compiler.core.common.cfg.ComputeBlockOrder;
 import org.graalvm.compiler.core.common.spi.ForeignCallSignature;
 import org.graalvm.compiler.core.common.spi.ForeignCallsProvider;
@@ -120,8 +121,8 @@ public abstract class Backend implements TargetProvider, ValueKindFactory<LIRKin
     /**
      * Creates a new instance of a block ordering computation.
      */
-    public ComputeBlockOrder newBlockOrder() {
-        return new BasicBlockOrder();
+    public <T extends AbstractBlockBase<T>> ComputeBlockOrder<T> newBlockOrder(int originalBlockCount, T startBlock) {
+        return new BasicBlockOrder<>(originalBlockCount, startBlock);
     }
 
     /**