+++ compiler/src/org.graalvm.compiler.printer/src/org/graalvm/compiler/printer/CFGPrinter.java
@@ -489,5 +489,4 @@ class CFGPrinter extends CompilationPrinter {
 
         printBlockEpilog(block);
     }
-
 }